move classes about
