+++ RepairsApi/V2/Helpers/IActivatorWrapper.cs
@@ -1,0 +1,7 @@
+namespace RepairsApi.V2.Helpers
+{
+    public interface IActivatorWrapper<T>
+    {
+        T CreateInstance<TConcrete>() where TConcrete : T;
+    }
+}
